requirements.txt
Django==4.0
django-extensions==3.1.5
djangorestframework==3.12.4
requests==2.27.1
pytz